third_party/blink/web_tests/external/wpt/css/css-variables/variable-definition-expected.txt

This is a testharness.js-based test.
Found 2 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] CSSOM.setProperty with space 1
  assert_equals: Expected Value should match actual value expected "" but got "green"
[FAIL] CSSOM.setProperty with space 2
  assert_equals: Expected Value should match actual value expected "" but got "green"
Harness: the test ran to completion.
